# docs-bose-api-spec.yaml
openapi: 3.0.3
info:
  title: Überböse Streaming API
  description: |
    The missing API specifications for your streaming device.

    This api spec combines the APIs of 4 different APIs / domains your speaker contacts.

    - **marge**: Account bound API managing source providers, presets, recents, and devices
    - **stats**: Device analytics and telemetry API collecting playback events and diagnostics
    - **software-update**: Firmware update information
    - **bmx**: API for streaming radio services (TuneIn, Custom Stations, etc.)
  version: 1.2.0
paths:
  /streaming/sourceproviders:
    get:
      summary: Get streaming source providers
      description: Retrieves a list of available streaming source providers
      operationId: getSourceProviders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response with list of source providers
          headers:
            Content-Type:
              schema:
                type: string
                example: application/vnd.bose.streaming-v1.2+xml
            ETag:
              schema:
                type: string
                example: "1762017390476"
          content:
            application/vnd.bose.streaming-v1.2+xml:
              schema:
                $ref: '#/components/schemas/SourceProvidersResponse'
              example: |
                <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
                <sourceProviders>
                  <sourceprovider id="1">
                    <createdOn>2012-09-19T12:43:00.000+00:00</createdOn>
                    <name>PANDORA</name>
                    <updatedOn>2012-09-19T12:43:00.000+00:00</updatedOn>
                  </sourceprovider>
                  <sourceprovider id="15">
                    <createdOn>2014-03-17T15:30:27.000+00:00</createdOn>
                    <name>SPOTIFY</name>
                    <updatedOn>2014-03-17T15:30:27.000+00:00</updatedOn>
                  </sourceprovider>
                </sourceProviders>
        '401':
          description: Unauthorized - Invalid or missing Bearer token
        '403':
          description: Forbidden - Valid token but insufficient permissions
        '500':
          description: Internal server error
  /streaming/account/{accountId}/device/{deviceId}/recent:
    post:
      summary: Add recent item
      description: Adds a recent item to the device's recent history
      operationId: addRecentItem
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/accountId'
        - $ref: '#/components/parameters/deviceId'
      requestBody:
        required: true
        content:
          application/vnd.bose.streaming-v1.2+xml:
            schema:
              $ref: '#/components/schemas/RecentItemRequest'
            example: |
              <?xml version="1.0" encoding="UTF-8" ?>
              <recent>
                <lastplayedat>2025-11-01T17:32:59+00:00</lastplayedat>
                <sourceid>19989313</sourceid>
                <name>Radio TEDDY</name>
                <location>/v1/playback/station/s80044</location>
                <contentItemType>stationurl</contentItemType>
              </recent>
      responses:
        '201':
          description: Recent item created successfully
          headers:
            Content-Type:
              schema:
                type: string
                example: application/vnd.bose.streaming-v1.2+xml
            Location:
              schema:
                type: string
                example: http://streamingqa.bose.com/account/6921042/device/587A628A4042/recent/2244536342
          content:
            application/vnd.bose.streaming-v1.2+xml:
              schema:
                $ref: '#/components/schemas/RecentItemResponse'
              example: |
                <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
                <recent id="2244536342">
                  <contentItemType>stationurl</contentItemType>
                  <createdOn>2018-11-27T18:20:01.000+00:00</createdOn>
                  <lastplayedat>2025-11-01T17:32:59.000+00:00</lastplayedat>
                  <location>/v1/playback/station/s80044</location>
                  <name>Radio TEDDY</name>
                  <source id="19989313" type="Audio">
                    <createdOn>2018-08-11T08:55:41.000+00:00</createdOn>
                    <credential type="token">eyDu=</credential>
                    <name></name>
                    <sourceproviderid>25</sourceproviderid>
                    <sourcename></sourcename>
                    <sourceSettings/>
                    <updatedOn>2019-07-20T17:48:31.000+00:00</updatedOn>
                    <username></username>
                  </source>
                  <sourceid>19989313</sourceid>
                  <updatedOn>2025-11-01T17:33:00.574+00:00</updatedOn>
                </recent>
        '400':
          description: Bad request - Invalid request body or parameters
        '401':
          description: Unauthorized - Invalid or missing Bearer token
        '403':
          description: Forbidden - Valid token but insufficient permissions
        '404':
          description: Not found - Account or device not found
        '500':
          description: Internal server error
  /streaming/account/{accountId}/device/{deviceId}/recents:
    get:
      summary: Get recent items
      description: Retrieves a list of recent items played on the device
      operationId: getRecents
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/accountId'
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Successful response with list of recent items
          headers:
            Content-Type:
              schema:
                type: string
                example: application/vnd.bose.streaming-v1.2+xml
            ETag:
              schema:
                type: string
                example: "1765724977209"
          content:
            application/vnd.bose.streaming-v1.2+xml:
              schema:
                $ref: '#/components/schemas/RecentsContainer'
              example: |
                <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
                <recents>
                  <recent id="2560855517">
                    <contentItemType>tracklisturl</contentItemType>
                    <createdOn>2025-12-13T17:14:28.000+00:00</createdOn>
                    <lastplayedat>2025-12-14T14:15:59.000+00:00</lastplayedat>
                    <location>/playback/container/c3BvdGlmeTphbGJ1bTowZ3BGWVZNbVV6VkVxeVAyeUh3cEha</location>
                    <name>Ghostsitter 23 - Das Haus im Moor</name>
                    <source id="19989621" type="Audio">
                      <createdOn>2018-08-11T09:52:31.000+00:00</createdOn>
                      <credential type="token_version_3">mockToken123=</credential>
                      <name>mockuser123</name>
                      <sourceproviderid>15</sourceproviderid>
                      <sourcename>mock@example.com</sourcename>
                      <sourceSettings/>
                      <updatedOn>2018-11-26T18:42:27.000+00:00</updatedOn>
                      <username>mockuser123</username>
                    </source>
                    <sourceid>19989621</sourceid>
                    <updatedOn>2025-12-14T14:16:05.000+00:00</updatedOn>
                  </recent>
                  <recent id="2244536335">
                    <contentItemType>stationurl</contentItemType>
                    <createdOn>2018-11-27T18:20:01.000+00:00</createdOn>
                    <lastplayedat>2025-12-14T12:32:00.000+00:00</lastplayedat>
                    <location>/v1/playback/station/s80044</location>
                    <name>Radio TEDDY</name>
                    <source id="19989313" type="Audio">
                      <createdOn>2018-08-11T08:55:41.000+00:00</createdOn>
                      <credential type="token">eyJmock=</credential>
                      <name></name>
                      <sourceproviderid>25</sourceproviderid>
                      <sourcename></sourcename>
                      <sourceSettings/>
                      <updatedOn>2019-07-20T17:48:31.000+00:00</updatedOn>
                      <username></username>
                    </source>
                    <sourceid>19989313</sourceid>
                    <updatedOn>2025-12-14T12:32:05.000+00:00</updatedOn>
                  </recent>
                </recents>
        '401':
          description: Unauthorized - Invalid or missing Bearer token
        '403':
          description: Forbidden - Valid token but insufficient permissions
        '404':
          description: Not found - Account or device not found
        '500':
          description: Internal server error
  /streaming/account/{accountId}/device/{deviceId}/recent/{recentId}:
    get:
      summary: Get recent item by ID
      description: Retrieves a specific recent item by its ID
      operationId: getRecentItem
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/accountId'
        - $ref: '#/components/parameters/deviceId'
        - name: recentId
          in: path
          required: true
          description: Recent item identifier
          schema:
            type: string
            example: "2557568761"
      responses:
        '200':
          description: Successful response with recent item details
          headers:
            Content-Type:
              schema:
                type: string
                example: application/vnd.bose.streaming-v1.2+xml
          content:
            application/vnd.bose.streaming-v1.2+xml:
              schema:
                $ref: '#/components/schemas/RecentItemResponse'
              example: |
                <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
                <recent id="3332">
                  <contentItemType>tracklisturl</contentItemType>
                  <createdOn>2025-11-08T21:07:34.000+00:00</createdOn>
                  <lastplayedat>2025-11-08T21:15:53.000+00:00</lastplayedat>
                  <location>/playback/container/c3BvdGlmeTp0cmFjazo0QnQ0VDRjM1c4UElrZEZuQVNLbXk5</location>
                  <name>Zähne putzen</name>
                  <source id="19989621" type="Audio">
                    <createdOn>2018-08-11T09:52:31.000+00:00</createdOn>
                    <credential type="token_version_3">mockToken789xyz=</credential>
                    <name>mockuser789xyz</name>
                    <sourceproviderid>15</sourceproviderid>
                    <sourcename>user@example.com</sourcename>
                    <sourceSettings/>
                    <updatedOn>2018-11-26T18:42:27.000+00:00</updatedOn>
                    <username>mockuser789xyz</username>
                  </source>
                  <sourceid>19989621</sourceid>
                  <updatedOn>2025-11-08T21:15:55.000+00:00</updatedOn>
                </recent>
        '401':
          description: Unauthorized - Invalid or missing Bearer token
        '403':
          description: Forbidden - Valid token but insufficient permissions
        '404':
          description: Not found - Account, device, or recent item not found
        '500':
          description: Internal server error
  /streaming/account/{accountId}/device/{deviceId}/presets:
    get:
      summary: Get device presets
      description: Retrieves a list of presets configured for the device
      operationId: getPresets
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/accountId'
        - $ref: '#/components/parameters/deviceId'
      responses:
        '200':
          description: Successful response with list of presets
          headers:
            Content-Type:
              schema:
                type: string
                example: application/vnd.bose.streaming-v1.2+xml
            ETag:
              schema:
                type: string
                example: "1736199076279"
          content:
            application/vnd.bose.streaming-v1.2+xml:
              schema:
                $ref: '#/components/schemas/PresetsContainer'
              example: |
                <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
                <presets>
                  <preset buttonNumber="1">
                    <containerArt>http://cdn-radiotime-logos.tunein.com/s80044q.png</containerArt>
                    <contentItemType>stationurl</contentItemType>
                    <createdOn>2018-11-26T18:40:45.000+00:00</createdOn>
                    <location>/v1/playback/station/s80044</location>
                    <name>Radio TEDDY</name>
                    <source id="19989313" type="Audio">
                      <createdOn>2018-08-11T08:55:41.000+00:00</createdOn>
                      <credential type="token">eyJzZXJpYWwiOiAiZjUwY2JjZGItN2IxMi00NjJmLWE5ZTctZGEzMTg5NDcwMTU4In0=</credential>
                      <name></name>
                      <sourceproviderid>25</sourceproviderid>
                      <sourcename></sourcename>
                      <sourceSettings/>
                      <updatedOn>2019-07-20T17:48:31.000+00:00</updatedOn>
                      <username></username>
                    </source>
                    <updatedOn>2018-11-26T18:40:45.000+00:00</updatedOn>
                    <username>Radio TEDDY</username>
                  </preset>
                  <preset buttonNumber="2">
                    <containerArt>https://mosaic.scdn.co/300/ab67616d0000b2732a4bc05d6d23b6a9a309e8f8ab67616d0000b2733d3c35cdbf9c51eefeaed7a7ab67616d0000b2738271a54c70aaae53e3da0c09ab67616d0000b273c4ec45569fc2db1edc44a465</containerArt>
                    <contentItemType>tracklisturl</contentItemType>
                    <createdOn>2018-11-14T18:27:39.000+00:00</createdOn>
                    <location>/playback/container/c3BvdGlmeTpwbGF5bGlzdDoyM1NNZHlPSEE2S2t6SG9QT0o1S1E5</location>
                    <name>Radio Mix</name>
                    <source id="19989621" type="Audio">
                      <createdOn>2018-08-11T09:52:31.000+00:00</createdOn>
                      <credential type="token_version_3">AQCfoupV5pdDJvqVjVdOF5tUOvyB3mKqeUGyZgqSYk4imOvta8p4dJnCHCNZANz-x0O-W2aQ15ML38pyx1CdvNSdOjGPwg0IlYYfPKWgoS07kmdEfRpf1Vur-OIx46KF56pjsA</credential>
                      <name>mock5zt8py3wuxy123xa431ge</name>
                      <sourceproviderid>15</sourceproviderid>
                      <sourcename>user@example.com</sourcename>
                      <sourceSettings/>
                      <updatedOn>2018-11-26T18:42:27.000+00:00</updatedOn>
                      <username>mock5zt8py3wuxy123xa431ge</username>
                    </source>
                    <updatedOn>2021-02-21T20:19:16.000+00:00</updatedOn>
                    <username>Radio Mix</username>
                  </preset>
                  <preset buttonNumber="3">
                    <containerArt>http://cdn-profiles.tunein.com/s25111/images/logoq.jpg?t=1</containerArt>
                    <contentItemType>stationurl</contentItemType>
                    <createdOn>2018-11-26T18:41:18.000+00:00</createdOn>
                    <location>/v1/playback/station/s25111</location>
                    <name>radioeins vom rbb</name>
                    <source id="19989313" type="Audio">
                      <createdOn>2018-08-11T08:55:41.000+00:00</createdOn>
                      <credential type="token">eyJzZXJpYWwiOiAiZjUwY2JjZGItN2IxMi00NjJmLWE5ZTctZGEzMTg5NDcwMTU4In0=</credential>
                      <name></name>
                      <sourceproviderid>25</sourceproviderid>
                      <sourcename></sourcename>
                      <sourceSettings/>
                      <updatedOn>2019-07-20T17:48:31.000+00:00</updatedOn>
                      <username></username>
                    </source>
                    <updatedOn>2018-11-26T18:41:18.000+00:00</updatedOn>
                    <username>radioeins vom rbb</username>
                  </preset>
                  <preset buttonNumber="4">
                    <containerArt>https://mosaic.scdn.co/300/ab67616d0000b2735ca387be0b44d742bcee3317ab67616d0000b273627072ee659b0f3aaa0f537aab67616d0000b27364b153e7f4b060d9735c7d76ab67616d0000b2738ba1aaf88b157453fe12482d</containerArt>
                    <contentItemType>tracklisturl</contentItemType>
                    <createdOn>2018-11-26T18:47:06.000+00:00</createdOn>
                    <location>/playback/container/c3BvdGlmeTpwbGF5bGlzdDoyd0JCOGIzUWhDWXd5T0d2dE9id3dI</location>
                    <name>Saisonal Simone Sommerland</name>
                    <source id="19989621" type="Audio">
                      <createdOn>2018-08-11T09:52:31.000+00:00</createdOn>
                      <credential type="token_version_3">AQCfoupV5pdDJvqVjVdOF5tUOvyB3mKqeUGyZgqSYk4imOvta8p4dJnCHCNZANz-x0O-W2aQ15ML38pyx1CdvNSdOjGPwg0IlYYfPKWgoS07kmdEfRpf1Vur-OIx46KF56pjsA</credential>
                      <name>mock5zt8py3wuxy123xa431ge</name>
                      <sourceproviderid>15</sourceproviderid>
                      <sourcename>user@example.com</sourcename>
                      <sourceSettings/>
                      <updatedOn>2018-11-26T18:42:27.000+00:00</updatedOn>
                      <username>mock5zt8py3wuxy123xa431ge</username>
                    </source>
                    <updatedOn>2022-11-17T19:35:37.000+00:00</updatedOn>
                    <username>Saisonal Simone Sommerland</username>
                  </preset>
                  <preset buttonNumber="5">
                    <containerArt></containerArt>
                    <contentItemType>tracklisturl</contentItemType>
                    <createdOn>2018-11-14T18:27:23.000+00:00</createdOn>
                    <location>/playback/container/c3BvdGlmeTpwbGF5bGlzdDozN2k5ZFFaRjFFNG1zazBURmdzeVRG</location>
                    <name>Ed Sheeran Radio</name>
                    <source id="20260226" type="Audio">
                      <createdOn>2018-09-16T19:18:46.000+00:00</createdOn>
                      <credential type="token_version_3">AQAHZ3A2USJzLX8CUEUPqIHOPpQihPpwcVO-0h2-g5-7SWRtIFfoQadeVu1Ud0TgVDkee7PyFnJbJUubZkrpv3A9jLaEeY991wSkXGG-EM_IlwvkRpASmwT3nbyixpVkSDw</credential>
                      <name>bldfvhhzdsyw2d4cy3melcaco</name>
                      <sourceproviderid>15</sourceproviderid>
                      <sourcename>otheruser@example.com</sourcename>
                      <sourceSettings/>
                      <updatedOn>2019-12-20T14:31:09.000+00:00</updatedOn>
                      <username>bldfvhhzdsyw2d4cy3melcaco</username>
                    </source>
                    <updatedOn>2020-01-26T12:45:13.000+00:00</updatedOn>
                    <username>Ed Sheeran Radio</username>
                  </preset>
                  <preset buttonNumber="6">
                    <containerArt>https://image-cdn-fa.spotifycdn.com/image/ab67706c0000da843b38733ef58fbd3530776a42</containerArt>
                    <contentItemType>tracklisturl</contentItemType>
                    <createdOn>2018-08-11T09:53:48.000+00:00</createdOn>
                    <location>/playback/container/c3BvdGlmeTpwbGF5bGlzdDoybjZXMnA1QzBNQUQ5YTR6NXhUVDdu</location>
                    <name>Komplett Entspannt</name>
                    <source id="19989621" type="Audio">
                      <createdOn>2018-08-11T09:52:31.000+00:00</createdOn>
                      <credential type="token_version_3">AQCfoupV5pdDJvqVjVdOF5tUOvyB3mKqeUGyZgqSYk4imOvta8p4dJnCHCNZANz-x0O-W2aQ15ML38pyx1CdvNSdOjGPwg0IlYYfPKWgoS07kmdEfRpf1Vur-OIx46KF56pjsA</credential>
                      <name>mock5zt8py3wuxy123xa431ge</name>
                      <sourceproviderid>15</sourceproviderid>
                      <sourcename>user@example.com</sourcename>
                      <sourceSettings/>
                      <updatedOn>2018-11-26T18:42:27.000+00:00</updatedOn>
                      <username>mock5zt8py3wuxy123xa431ge</username>
                    </source>
                    <updatedOn>2025-01-06T21:31:15.000+00:00</updatedOn>
                    <username>Komplett Entspannt</username>
                  </preset>
                </presets>
        '401':
          description: Unauthorized - Invalid or missing Bearer token
        '403':
          description: Forbidden - Valid token but insufficient permissions
        '404':
          description: Not found - Account or device not found
        '500':
          description: Internal server error
...
  /streaming/account/{accountId}/device/:
    post:
      summary: Register device to account
      description: |
        Registers/pairs a device (speaker) to a marge account.
        Associates the device ID with the marge account ID.
        This can be triggerd by calling:

        ```
        POST http://<speaker-id>:8090/setMargeAccount
        Accept: application/xml
        Content-Type: application/xml

        <PairDeviceWithAccount>
          <accountId>6921042</accountId>
          <userAuthToken>Bearer test123</userAuthToken>
        </PairDeviceWithAccount>
        ```
      operationId: addDevice
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/accountId'
      requestBody:
        required: true
        content:
          application/vnd.bose.streaming-v1.2+xml:
            schema:
              $ref: '#/components/schemas/DeviceUpdateRequest'
            example: |
              <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
              <device deviceid="587A628A4042">
                <name>Kitchen</name>
                <macaddress>587A628A4042</macaddress>
              </device>
      responses:
        '201':
          description: Device registered successfully
          headers:
            Content-Type:
              schema:
                type: string
                example: application/vnd.bose.streaming-v1.2+xml
            Credentials:
              schema:
                type: string
                example: Bearer nRW/F+ads/3fsfd//sdfs//fsdfs/qj/wer
            Location:
              schema:
                type: string
                example: http://streamingqa.bose.com/account/6921042/device/587A628A4042
            METHOD_NAME:
              schema:
                type: string
                example: addDevice
          content:
            application/vnd.bose.streaming-v1.2+xml:
              schema:
                $ref: '#/components/schemas/DeviceUpdateResponse'
              example: |
                <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
                <device deviceid="587A628A4042">
                  <createdOn>2018-08-11T08:55:25.000+00:00</createdOn>
                  <ipaddress></ipaddress>
                  <name>Kitchen</name>
                  <updatedOn>2026-01-02T11:12:09.074+00:00</updatedOn>
                </device>
        '401':
          description: Unauthorized - Invalid or missing Bearer token
          content:
            application/vnd.bose.streaming-v1.2+xml:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: |
                <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
                <status>
                  <message>Not Authorized</message>
                  <status-code>401</status-code>
                </status>
...
  /bmx/registry/v1/services:
    get:
      summary: Get BMX services registry
      description: |
        Returns the list of available BMX streaming services (TuneIn, Custom Stations, SiriusXM, Radioplayer).
        This endpoint provides service discovery for streaming audio providers.
      operationId: getBmxServices
      tags:
        - bmx
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful response with list of BMX services
          headers:
            Content-Type:
              schema:
                type: string
                example: application/json
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BmxServicesResponse'
        '401':
          description: Unauthorized - Invalid or missing Bearer token
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer token authentication
  parameters:
    accountId:
      name: accountId
      in: path
      required: true
      description: Account identifier
      schema:
        type: string
        pattern: '^[0-9a-zA-Z-]+$'
        example: "6921042"
    deviceId:
      name: deviceId
      in: path
      required: true
      description: Device identifier
      schema:
        type: string
        example: "587A628A4042"
  schemas:
    SourceProvidersResponse:
      type: object
      description: XML response containing list of source providers
      xml:
        name: sourceProviders
      properties:
        sourceprovider:
          type: array
          items:
            $ref: '#/components/schemas/SourceProvider'
          xml:
            wrapped: false
...
    PresetsContainer:
      type: object
      description: Container for presets
      xml:
        name: presets
      properties:
        preset:
          type: array
          items:
            $ref: '#/components/schemas/Preset'
          xml:
            wrapped: false
...
    RecentsContainer:
      type: object
      description: Container for recent items
      xml:
        name: recents
      properties:
        recent:
          type: array
          items:
            $ref: '#/components/schemas/RecentItem'
          xml:
            wrapped: false
...